// src/common/types/types.ts
export enum UserRoles {
    user = 'user',
    admin = 'admin'
}

export enum UserGender {
    male = 'male',
    female = 'female'
}

export enum OTPTypes {
    confirmation = 'confirmation',
    resetPassword = 'resetPassword'
}

export enum PaymentMethodTypes {
    cash = 'cash',
    card = 'card'
}

export enum OrderStatusTypes {
    pending = 'pending',
    placed = 'placed',
    onWay = 'onWay',
    delivered = 'delivered',
    cancelled = 'cancelled',
    rejected = 'rejected',
    refunded = 'refunded',
    paid = 'paid'
}